Fix NoClassDefFoundErrors on 1.21.5
Try/catch containing new classes gets evaluated by the JVM on class load.
