+++ internal/common/src/main/java/com/lishid/openinv/internal/common/container/slot/ContentCursor.java
@@ -1,7 +1,7 @@
 package com.lishid.openinv.internal.common.container.slot;
 
 import com.lishid.openinv.internal.common.container.slot.placeholder.Placeholders;
-import com.lishid.openinv.internal.common.player.OpenPlayer;
+import com.lishid.openinv.internal.common.player.BaseOpenPlayer;
 import net.minecraft.server.level.ServerPlayer;
 import net.minecraft.world.Container;
 import net.minecraft.world.entity.player.Player;
@@ -64,7 +64,7 @@ public class ContentCursor implements Content {
     // Player must be online and not in creative - since the creative client is (semi-)authoritative,
     // it ignores changes without extra help, and will delete the item as a result.
     // Spectator mode is technically possible but may cause the item to be dropped if the client opens an inventory.
-    return OpenPlayer.isConnected(holder.connection) && holder.gameMode.isSurvival();
+    return BaseOpenPlayer.isConnected(holder.connection) && holder.gameMode.isSurvival();
   }
 
   @Override
