+++ internal/paper1_21_5/src/main/java/com/lishid/openinv/internal/paper1_21_5/player/OpenPlayer.java
@@ -1,6 +1,7 @@
 package com.lishid.openinv.internal.paper1_21_5.player;
 
 import com.lishid.openinv.event.OpenEvents;
+import com.lishid.openinv.internal.common.player.BaseOpenPlayer;
 import com.mojang.logging.LogUtils;
 import net.minecraft.Util;
 import net.minecraft.nbt.CompoundTag;
@@ -8,12 +9,11 @@ import net.minecraft.nbt.NbtIo;
 import net.minecraft.server.level.ServerPlayer;
 import net.minecraft.world.level.storage.PlayerDataStorage;
 import org.bukkit.craftbukkit.CraftServer;
-import org.slf4j.Logger;
 
 import java.nio.file.Files;
 import java.nio.file.Path;
 
-public class OpenPlayer extends com.lishid.openinv.internal.common.player.OpenPlayer {
+public class OpenPlayer extends BaseOpenPlayer {
 
   protected OpenPlayer(CraftServer server, ServerPlayer entity, PlayerManager manager) {
     super(server, entity, manager);
@@ -26,7 +26,6 @@ public class OpenPlayer extends com.lishid.openinv.internal.common.player.OpenPl
     }
 
     ServerPlayer player = this.getHandle();
-    Logger logger = LogUtils.getLogger();
     // See net.minecraft.world.level.storage.PlayerDataStorage#save(EntityHuman)
     try {
       PlayerDataStorage worldNBTStorage = player.server.getPlayerList().playerIo;
